+++ rust/candid/tests/parse_type.rs
@@ -1,6 +1,7 @@
 use candid::bindings::{candid as candid_export, javascript, motoko, rust, typescript};
-use candid::parser::types::{to_pretty, IDLProg};
-use candid::parser::typing::{check_file, check_prog, TypeEnv};
+use candid::parser::types::IDLProg;
+use candid::parser::typing::{check_file, check_prog};
+use candid::types::TypeEnv;
 use goldenfile::Mint;
 use std::io::Write;
 use std::path::Path;
@@ -25,10 +26,7 @@ service server : {
   i : f;
 }
     "#;
-    let ast = prog.parse::<IDLProg>().unwrap();
-    let pretty = to_pretty(&ast, 80);
-    let ast2 = pretty.parse::<IDLProg>().unwrap();
-    assert_eq!(format!("{ast2:?}"), format!("{ast:?}"));
+    prog.parse::<IDLProg>().unwrap();
 }
 
 #[test_generator::test_resources("rust/candid/tests/assets/*.did")]