+++ rust/candid/tests/serde.rs
@@ -1,4 +1,3 @@
-#![allow(clippy::unit_cmp)]
 use candid::{decode_one, encode_one, CandidType, Decode, Deserialize, Encode, Int, Nat};
 
 #[test]